refactor : data class model 폴더로 이동
