+++ WssElixir/app/src/main/java/com/example/wsselixir/presentation/model/User.kt
@@ -1,4 +1,4 @@
-package com.example.wsselixir.presentation.home
+package com.example.wsselixir.presentation.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
